Update calculateTrades.m
I updated the trading code to accurately reflect what the professor described in class. The last code identified a signal and then traded in the NEXT bar. Now the code identifies a signal and trades at the current bar either at the breakout point (the running max or min) or at the open (if there was a gap in the price).

I also updated the stpPct part of the trading strategy. Before I was only updating the peak/trough based on each bar's open, but now it updates based on the high/low. And like above, I took care of the possibility of a gap.